reorganize payload handlers, split out dcppc into separate handler
