+++ src/payload_handlers/__init__.py
@@ -3,15 +3,15 @@ from .factory import PayloadHandlerFactory
 from .handlers import \
         BasePayloadHandler, \
         PRTestingPayloadHandler, \
-        MCTestingPayloadHandler, \
-        LoggingPayloadHandler, \
+        LoggingPayloadHandler
+
+from .dcppc import \
         DCPPCPayloadHandler
 
 __all__ = [
         'PayloadHandlerFactory',
         'BasePayloadHandler',
         'PRTestingPayloadHandler',
-        'MCTestingPayloadHandler',
         'LoggingPayloadHandler',
         'DCPPCPayloadHandler'
 ]
